Set Role and permission middleware on Doctor

+++ app/Http/Controllers/Backsite/DoctorController.php
@@ -3,8 +3,9 @@
 namespace App\Http\Controllers\Backsite;
 
 use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
 
+use Illuminate\Support\Str;
+use Illuminate\Http\Request;
 Use Illuminate\Support\Facades\Storage;
 Use Symfony\Component\HttpFoundation\Response;
 
@@ -12,7 +13,11 @@ Use Symfony\Component\HttpFoundation\Response;
 Use App\Http\Requests\Doctor\StoreDoctorRequest;
 Use App\Http\Requests\Doctor\UpdateDoctorRequest;
 
+// Use everything here
+Use Gate;
 Use Auth;
+
+// Models
 Use App\Models\Operational\Doctor;
 Use App\Models\MasterData\Specialist;
 
@@ -29,6 +34,8 @@ class DoctorController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('doctor_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $doctor = Doctor::orderBy('created_at', 'desc')->get();
 
         $specialist = Specialist::orderBy('name', 'asc')->get();
@@ -70,6 +77,9 @@ class DoctorController extends Controller
      */
     public function show(Doctor $doctor)
     {
+        // Abort if the requested doesn't have permission by user role
+        abort_if(Gate::denies('doctor_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('pages.backsite.operational.doctor.show', compact('doctor'));
     }
 
@@ -81,6 +91,9 @@ class DoctorController extends Controller
      */
     public function edit(Doctor $doctor)
     {
+        // Abort if the requested doesn't have permission by user role
+        abort_if(Gate::denies('doctor_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $specialist = Specialist::orderBy('name', 'asc')->get();
 
         return view('pages.backsite.operational.doctor.edit', compact('doctor', 'specialist'));
@@ -111,6 +124,9 @@ class DoctorController extends Controller
      */
     public function destroy(Doctor $doctor)
     {
+        // Abort if the requested doesn't have permission by user role
+        abort_if(Gate::denies('doctor_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $doctor->forceDelete();
 
         alert()->success('Success', 'Your Doctor has been deleted');
